+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.7.3dev
 * [Doc] Tutorial on querying excel files with pandas and jupysql ([#423](https://github.com/ploomber/jupysql/pull/423))
 * [Fix] Fix `--alias` when passing an existing engine
-* [Fix] Fix displaylimit default value to 60 (#462)
+* [Fix] Fix displaylimit default value to 10 (#462)
 
 ## 0.7.2 (2023-04-25)